refining repl communication; refining series operations
